+++ server/middlewares/index.js
@@ -1,4 +1,5 @@
 const isAuth = require("./isAuth");
 const secure = require("./secure");
 const { errorConvert, globalErrorHandler } = require("./error");
+
 module.exports = { isAuth, secure, errorConvert, globalErrorHandler };
